Remove the "increase" speed up to avoid confusion in 1-7.txt
The "increase" in speed up 40% and 51.33% are wrong.

The simple speed up is defined as the ratio of the execution times, that are 1.67 and 2.27.

The "increase" in speed up should refer to how much over the base case (0.66 seconds) the new time
has grown.

For the case of compiler A: A 67% increase over 0.66 has occurred -> 0.67*0.66 + 0.66 = (1+0.67)*0.66 = 1.1
For the case of compiler B: A 127% increase over 0.66 has occurred -> 1.27*0.66 + 0.66 = (1+1.27)*0.66 = 1.5